Modified the firmware and comm for HW v0.2
The HackSense hardware version 0.2 uses a four-digit seven-segment LED
display instead of 2 LEDs as a feedback to the user, so the firmware had
to be modified to handle the new peripheal. The serial protocol also
needed a few changes to be able to handle states 0-99 (compared to 0-1
in v0.1), so the communication component was altered accordingly.